+++ batcher/aligned-batcher/src/connection.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::types::{batch_queue::BatchQueueEntry, errors::BatcherError};
 use aligned_sdk::{
     communication::serialization::cbor_serialize,
-    core::types::{BatchInclusionData, SubmitProofResponseMessage, VerificationCommitmentBatch},
+    common::types::{BatchInclusionData, SubmitProofResponseMessage, VerificationCommitmentBatch},
 };
 use futures_util::{stream::SplitSink, SinkExt};
 use lambdaworks_crypto::merkle_tree::merkle::MerkleTree;
