+++ batcher/aligned-batcher/src/eth/utils.rs
@@ -8,7 +8,7 @@ use crate::{
         retry_function,
     },
 };
-use aligned_sdk::core::constants::{
+use aligned_sdk::common::constants::{
     ETHEREUM_CALL_BACKOFF_FACTOR, ETHEREUM_CALL_MAX_RETRIES, ETHEREUM_CALL_MAX_RETRY_DELAY,
     ETHEREUM_CALL_MIN_RETRY_DELAY, GAS_PRICE_INCREMENT_PERCENTAGE_PER_ITERATION,
     OVERRIDE_GAS_PRICE_PERCENTAGE_MULTIPLIER, PERCENTAGE_DIVIDER,
