+++ batcher/aligned-sdk/src/common/errors.rs
