+++ batcher/aligned-sdk/src/communication/batch.rs
@@ -1,14 +1,13 @@
 use log::debug;
 
 use crate::{
-    core::{
+    common::{
         errors,
         types::{
             AlignedVerificationData, BatchInclusionData, Network, VerificationCommitmentBatch,
             VerificationDataCommitment,
         },
-    },
-    sdk::is_proof_verified,
+    }, verification_layer::is_proof_verified
 };
 
 const RETRIES: u64 = 10;
