+++ batcher/aligned-sdk/src/communication/messaging.rs
@@ -11,10 +11,10 @@ use futures_util::stream::{SplitSink, TryFilter};
 use tokio_tungstenite::{tungstenite::Message, MaybeTlsStream, WebSocketStream};
 
 use crate::communication::serialization::{cbor_deserialize, cbor_serialize};
-use crate::core::types::{BatchInclusionData, SubmitProofMessage};
+use crate::common::types::{BatchInclusionData, SubmitProofMessage};
 use crate::{
     communication::batch::process_batcher_response,
-    core::{
+    common::{
         errors::SubmitError,
         types::{
             AlignedVerificationData, ClientMessage, NoncedVerificationData,
