+++ batcher/aligned-sdk/src/eth/aligned_proof_agg_service.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use ethers::prelude::*;
 
-use crate::core::errors::VerificationError;
+use crate::common::errors::VerificationError;
 
 abigen!(
     ProofAggregationServiceContract,
